# pq_app/api/openapi.yaml
openapi: 3.0.0
info:
  title: PyQuest API
  description: RESTful API for PyQuest game - mobile and external client access
  version: 1.0.0
  contact:
    name: PyQuest Team
    email: support@pyquest.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.pyquest.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CombatAction:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
          example: "attack_light"
        name:
          type: string
          example: "Light Attack"
        description:
          type: string
        damage_min:
          type: integer
          example: 3
        damage_max:
          type: integer
          example: 8
        heal_amount:
          type: integer
          example: 0
        defense_boost:
          type: integer
          example: 0
        success_rate:
          type: integer
          example: 95
        requires_class:
          type: string
          nullable: true
        requires_race:
          type: string
          nullable: true
    Encounter:
      type: object
      properties:
        id:
          type: integer
        player_id:
          type: integer
        tile_id:
          type: integer
        combat_action_id:
          type: integer
        player_hp_before:
          type: integer
        player_hp_after:
          type: integer
        damage_dealt:
          type: integer
        damage_received:
          type: integer
        was_successful:
          type: boolean
        created_at:
          type: string
          format: date-time
paths:
...
  /player/{player_id}/tiles/{tile_id}/combat-actions:
    get:
      tags:
        - Combat
      summary: Get available combat actions for tile
      security:
        - bearerAuth: []
      parameters:
        - name: player_id
          in: path
          required: true
          schema:
            type: integer
        - name: tile_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Available combat actions
          content:
            application/json:
              schema:
                type: object
                properties:
                  tile_id:
                    type: integer
                  tile_type:
                    type: integer
                  available_actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/CombatAction'
  /player/{player_id}/combat/execute:
    post:
      tags:
        - Combat
      summary: Execute a combat action
      security:
        - bearerAuth: []
      parameters:
        - name: player_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tile_id
                - combat_action_code
              properties:
                tile_id:
                  type: integer
                combat_action_code:
                  type: string
                  example: "attack_heavy"
      responses:
        '200':
          description: Combat result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  player_hp:
                    type: integer
                  damage_dealt:
                    type: integer
                  damage_received:
                    type: integer
                  gold_change:
                    type: integer
                  experience_change:
                    type: integer
                  encounter:
                    $ref: '#/components/schemas/Encounter'
  /player/{player_id}/encounters:
    get:
      tags:
        - Combat
      summary: Get encounter history
      security:
        - bearerAuth: []
      parameters:
        - name: player_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of encounters
          content:
            application/json:
              schema:
                type: object
                properties:
                  encounters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Encounter'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer